Add files via upload
This is the original file (for the purposes of this stage of the project) which Prajnamati has cleaned up and uploaded into a Table Viewer which can be used with a Wordpress website.  It's not complete yet as it doesn't have Bhante's Guhyaloka books on it, nor the new books of his which have yet to be added to the catalogue.