reformat of SL DhBks Main v9. | separated values
Reformat of SL DhBks Main v9 to replace 3 existing "|" characters with spaces, then replace all "," with "|".
This is a text file (.txt) containing vertical bar-separated values. It can be imported in to Excel.
A text file allows the edit and compare functions on GitHub. 
Using | as a separator allows for the use of commas in all fields - so we can reproduce the commas in titles